# be/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PlanWise API Documentation",
        "description": "API Documentation for PlanWise application - a smart daily planning application with AI features.",
        "contact": {
            "name": "PlanWise Team",
            "email": "contact@planwise.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/api/v1/goals/{goalId}": {
            "get": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "getGoalById",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "updateGoal",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGoalRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "deleteGoal",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "updateGoalProgress",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGoalRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/goals/{goalId}/milestones/{milestoneId}": {
            "put": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "updateMilestone",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    {
                        "name": "milestoneId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMilestoneRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MilestoneDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "deleteMilestone",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    {
                        "name": "milestoneId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/goals": {
            "get": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "getGoals",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "week",
                                "month",
                                "year"
                            ],
                            "example": "week"
                        },
                        "example": "week"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GoalDto"
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/GoalListResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "createGoal",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGoalRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GoalDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/goals/{goalId}/milestones": {
            "get": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "getMilestones",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MilestoneDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "goal-controller"
                ],
                "operationId": "createMilestone",
                "parameters": [
                    {
                        "name": "goalId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                        },
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMilestoneRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/MilestoneDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateGoalRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "career",
                            "learning",
                            "health",
                            "finance"
                        ],
                        "example": "career"
                    },
                    "goalType": {
                        "type": "string",
                        "enum": [
                            "SMART",
                            "OKR"
                        ],
                        "example": "SMART"
                    },
                    "period": {
                        "type": "string",
                        "enum": [
                            "week",
                            "month",
                            "year"
                        ],
                        "example": "week"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "progress": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "isCompleted": {
                        "type": "boolean"
                    }
                }
            },
            "GoalDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "career",
                            "learning",
                            "health",
                            "finance"
                        ],
                        "example": "career"
                    },
                    "goalType": {
                        "type": "string",
                        "enum": [
                            "SMART",
                            "OKR"
                        ],
                        "example": "SMART"
                    },
                    "period": {
                        "type": "string",
                        "enum": [
                            "week",
                            "month",
                            "year"
                        ],
                        "example": "week"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "progress": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    },
                    "isCompleted": {
                        "type": "boolean"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "milestones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MilestoneDto"
                        }
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
            "MilestoneDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-06-05T08:00:00.000Z"
                    }
                }
            },
            "UpdateMilestoneRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "completed": {
                        "type": "boolean"
                    }
                }
            },
...
            "CreateGoalRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "career",
                            "learning",
                            "health",
                            "finance"
                        ],
                        "example": "career"
                    },
                    "goalType": {
                        "type": "string",
                        "enum": [
                            "SMART",
                            "OKR"
                        ],
                        "example": "SMART"
                    },
                    "period": {
                        "type": "string",
                        "enum": [
                            "week",
                            "month",
                            "year"
                        ],
                        "example": "week"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    },
                    "color": {
                        "type": "string",
                        "example": "indigo"
                    }
                },
                "required": [
                    "period",
                    "title"
                ]
            },
            "CreateMilestoneRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "targetDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-05"
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "GoalListResponse": {
                "type": "object",
                "properties": {
                    "goals": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GoalDto"
                        }
                    },
                    "totalCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "weeklyCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "monthlyCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "yearlyCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isPremium": {
                        "type": "boolean"
                    },
                    "freeLimit": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter your Bearer JWT token to access protected endpoints.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}